chat/requirements.txt
groq>=0.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
windows-curses>=2.3.1; platform_system == "Windows"